# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: YARB API
  version: '1.0'
servers:
  - url: 'http://localhost:3000'
paths:
  /auth/login:
    post:
      summary: login
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: ./models/user/login-data.v1.yaml
        '400':
          description: 'Bad Request: Invalid input'
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: ./models/internal-error-message.v1.yaml
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./models/user/user-credentials.v1.yaml
        description: ''
      description: Create token by user credentials and get userId
      operationId: login
  /boards:
    get:
      summary: getBoards
      tags: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: ./models/board/board.v1.yaml
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: ./models/internal-error-message.v1.yaml
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
          in: query
          name: userId
          required: true
      operationId: getBoards
      description: Get boards by owner
    post:
      summary: createBoard
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: ./models/created-response.v1.yaml
        '401':
          description: Unauthorized
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./models/board/create-board.v1.yaml
      description: Create new board
      operationId: createBoard
  /users:
    post:
      summary: createUser
      responses:
        '204':
          description: No Content
        '409':
          description: "Conflict: username already existing          \n"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: ./models/internal-error-message.v1.yaml
      description: Create new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./models/user/user-credentials.v1.yaml
      operationId: createUser
    parameters: []
  '/users/{userId}':
    parameters:
      - schema:
          type: integer
        name: userId
        in: path
        required: true
    get:
      summary: getUser
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: ./models/user/user.v1.yaml
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: ./models/internal-error-message.v1.yaml
      description: Get user
      operationId: getUser
      security:
        - bearerAuth: []
components:
  schemas: {}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: Bearer